리뷰 이미지 관련 에러 수정

+++ src/main/java/com/multi/multi_semi/review/dao/ReviewMapper.java
@@ -18,7 +18,7 @@ public interface ReviewMapper {
 
     int selectReviewTotal();
 
-    ReviewResDto findReviewByNo(@Param("no")int reviewNo);
+    ReviewResDto findReviewByNo(@Param("no")Long reviewNo);
 
     int insertReview(ReviewReqDto reviewReqDto);
